+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/clients/gemini/GeminiClientService.kt
@@ -4,7 +4,9 @@ import com.github.blarc.ai.commits.intellij.plugin.settings.clients.LLMClientSer
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import dev.langchain4j.model.chat.ChatLanguageModel
+import dev.langchain4j.model.chat.StreamingChatLanguageModel
 import dev.langchain4j.model.vertexai.VertexAiGeminiChatModel
+import dev.langchain4j.model.vertexai.VertexAiGeminiStreamingChatModel
 import kotlinx.coroutines.CoroutineScope
 
 @Service(Service.Level.APP)
@@ -24,4 +26,12 @@ class GeminiClientService(private val cs: CoroutineScope): LLMClientService<Gemi
             .build()
     }
 
+    override suspend fun buildStreamingChatModel(client: GeminiClientConfiguration): StreamingChatLanguageModel {
+        return VertexAiGeminiStreamingChatModel.builder()
+            .project(client.projectId)
+            .location(client.location)
+            .modelName(client.modelId)
+            .temperature(client.temperature.toFloat())
+            .build()
+    }
 }